Refactor : div component 폴더 생성 및 파일 위치 이동 - div component를 별도로 관리하기 위한 폴더를 생성함
