+++ Client/src/pages/milestone/MilestonePage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Container from '../../components/Container';
+import Container from '../../components/div/Container';
 import Menu from '../../components/Menu';
 
 export default function MilestonePage() {
